# Conflicts:
#	src/main/java/org/example/cinema_reservation_system/entity/DaoDien.java
#	src/main/java/org/example/cinema_reservation_system/entity/DienVien.java
#	src/main/java/org/example/cinema_reservation_system/entity/HinhAnh.java
#	src/main/java/org/example/cinema_reservation_system/entity/PhongChieu.java
#	src/main/java/org/example/cinema_reservation_system/entity/RapChieu.java
#	src/main/java/org/example/cinema_reservation_system/entity/Role.java
#	src/main/java/org/example/cinema_reservation_system/entity/TheLoai.java
#	src/main/java/org/example/cinema_reservation_system/entity/UserAccount.java
#	src/main/resources/application.properties

+++ src/main/java/org/example/cinema_reservation_system/dto/phongchieudto/PhongChieuResponseDto.java
@@ -5,13 +5,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class PhongChieuResponseDto {//trả dữ liệu về FE
     private Integer idPhongChieu;
     private String tenPhongChieu;
-    private Double dienTichPhong;
+    private BigDecimal dienTichPhong;
     private Enum.TrangThaiPhongChieu trangThaiPhongChieu;
 
     // Thông tin từ Rạp Chiếu
